moved AccountHelper and created an interface for it
(better testing)
